appEngine/requirements.txt
Flask==0.12.2
Pillow==5.1.0
Werkzeug<0.13.0,>=0.12.0
google-cloud-storage==1.6.0
google-api-python-client==1.7.3
gunicorn==19.7.1
